+++ src/main/java/com/omelentjeff/citybike/repository/JourneyRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface JourneyRepository extends JpaRepository<Journey, Integer>  {
+public interface JourneyRepository extends JpaRepository<Journey, Long>  {
 }
